+++ web/src/main/java/com/homeapp/javatraining/exception/question/NotEnoughQuestionsException.java
@@ -1,0 +1,10 @@
+package com.homeapp.javatraining.exception.question;
+
+import com.homeapp.javatraining.exception.ApiException;
+
+public class NotEnoughQuestionsException extends ApiException {
+
+    public NotEnoughQuestionsException(int requested, int available) {
+        super("Requested " + requested + " questions, but only " + available + " are available");
+    }
+}