+++ web/src/main/java/com/homeapp/javatraining/exception/question/QuestionImportException.java
@@ -1,0 +1,10 @@
+package com.homeapp.javatraining.exception.question;
+
+import com.homeapp.javatraining.exception.ApiException;
+
+public class QuestionImportException extends ApiException {
+
+    public QuestionImportException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}