+++ web/src/main/java/com/homeapp/javatraining/exception/question/QuestionNotFoundException.java
@@ -1,0 +1,10 @@
+package com.homeapp.javatraining.exception.question;
+
+import com.homeapp.javatraining.exception.ApiException;
+
+public class QuestionNotFoundException extends ApiException {
+    public QuestionNotFoundException(Long questionId) {
+        super("Question with id " + questionId + " not found");
+    }
+
+}