+++ web/src/main/java/com/homeapp/javatraining/exception/user/CannotChangeOwnRoleException.java
@@ -1,0 +1,10 @@
+package com.homeapp.javatraining.exception.user;
+
+import com.homeapp.javatraining.exception.ApiException;
+
+public class CannotChangeOwnRoleException extends ApiException {
+
+    public CannotChangeOwnRoleException() {
+        super("You cannot change your own role");
+    }
+}