+++ web/src/main/java/com/homeapp/javatraining/exception/user/DuplicateUsernameException.java
@@ -1,0 +1,10 @@
+package com.homeapp.javatraining.exception.user;
+
+import com.homeapp.javatraining.exception.ApiException;
+
+public class DuplicateUsernameException extends ApiException {
+
+    public DuplicateUsernameException(String username) {
+        super("Username '" + username + "' is already in use");
+    }
+}