+++ web/src/main/java/com/homeapp/javatraining/service/AdminUserService.java
@@ -1,43 +1,48 @@
 package com.homeapp.javatraining.service;
 
+import com.homeapp.javatraining.exception.user.CannotChangeOwnRoleException;
+import com.homeapp.javatraining.exception.user.UserNotFoundException;
 import com.homeapp.javatraining.model.Role;
+import com.homeapp.javatraining.model.User;
 import com.homeapp.javatraining.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class AdminUserService {
-    private static final Logger log = LoggerFactory.getLogger(AdminUserService.class);
 
     private final UserRepository userRepository;
 
-
     public void changeUserRole(long adminId, long targetUserId, Role newRole) {
+
         log.info("Admin {} attempts to change role of user {} to {}",
                 adminId,
                 targetUserId,
-                newRole
-        );
+                newRole);
 
-        userRepository.findById(targetUserId).ifPresentOrElse(user -> {
-            if (user.getId() == adminId) {
-                log.warn("Admin {} attempted to change own role. Operation denied", adminId);
-                return;
-            }
-            user.setRole(newRole);
-            userRepository.save(user);
-            log.info("User {} role changed to {} by admin {}",
-                    targetUserId,
-                    newRole,
-                    adminId
-            );
-        }, () -> {
-            log.warn("Admin {} attempted to change role of non-existing user {}",
-                    adminId,
-                    targetUserId);
-        });
+        User user = userRepository.findById(targetUserId)
+                .orElseThrow(() -> {
+                    log.warn("Admin {} attempted to change role of non-existing user {}",
+                            adminId,
+                            targetUserId);
+                    return new UserNotFoundException(targetUserId);
+                });
+
+        if (user.getId() == adminId) {
+            log.warn("Admin {} attempted to change own role", adminId);
+            throw new CannotChangeOwnRoleException();
+        }
+
+        user.setRole(newRole);
+        userRepository.save(user);
+
+        log.info("User {} role changed to {} by admin {}",
+                targetUserId,
+                newRole,
+                adminId);
     }
 }